requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
playwright==1.58.0
pyee==13.0.0
python-dotenv==1.2.1
python-telegram-bot==21.0.1
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
